Issue-212: Moved PaginatedScreenVM.kt and PaginatedScreenFragment.kt to core module
