+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/tvs/TvsFragment.kt
@@ -1,10 +1,10 @@
 package com.kirchhoff.movies.ui.screens.main.tvs
 
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenFragment
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.data.ui.main.UITv
-import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.core.tvs.adapter.TvsListAdapter
 import com.kirchhoff.movies.ui.screens.details.DetailsActivity
 import org.koin.android.viewmodel.ext.android.viewModel
