+++ app/src/main/java/com/kirchhoff/movies/ui/screens/reviews/list/ReviewsListFragment.kt
@@ -3,10 +3,10 @@ package com.kirchhoff.movies.ui.screens.reviews.list
 import android.os.Bundle
 import android.view.View
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenFragment
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.data.ui.details.review.UIReview
-import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.reviews.ReviewType
 import com.kirchhoff.movies.ui.screens.reviews.details.ReviewDetailsFragment
 import com.kirchhoff.movies.ui.screens.reviews.list.adapter.ReviewsListAdapter
