+++ app/src/main/java/com/kirchhoff/movies/ui/screens/reviews/list/ReviewsListVM.kt
@@ -1,10 +1,10 @@
 package com.kirchhoff.movies.ui.screens.reviews.list
 
+import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenVM
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.details.review.UIReview
 import com.kirchhoff.movies.repository.movie.IMovieRepository
 import com.kirchhoff.movies.repository.tv.ITvRepository
-import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
 import com.kirchhoff.movies.ui.screens.reviews.ReviewType
 
 class ReviewsListVM(
