+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/tv/SimilarTvsVM.kt
@@ -1,9 +1,9 @@
 package com.kirchhoff.movies.ui.screens.similar.tv
 
+import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenVM
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.data.ui.main.UITv
 import com.kirchhoff.movies.repository.tv.ITvRepository
-import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
 
 class SimilarTvsVM(
     private val tvId: Int,
