+++ core/src/main/java/com/kirchhoff/movies/core/ui/paginated/PaginatedScreenFragment.kt
@@ -1,4 +1,4 @@
-package com.kirchhoff.movies.ui.screens.core
+package com.kirchhoff.movies.core.ui.paginated
 
 import android.os.Bundle
 import android.view.View
@@ -6,15 +6,14 @@ import android.widget.Toast
 import androidx.annotation.StringRes
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.GridLayoutManager
-import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.R
+import com.kirchhoff.movies.core.databinding.FragmentPaginatedBinding
 import com.kirchhoff.movies.core.ui.BaseFragment
-import com.kirchhoff.movies.core.ui.paginated.UIPaginated
 import com.kirchhoff.movies.core.ui.recyclerview.Paginator
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.BaseRecyclerViewAdapter
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
 import com.kirchhoff.movies.core.ui.recyclerview.decorations.GridMarginItemDecoration
 import com.kirchhoff.movies.core.ui.utils.viewBinding
-import com.kirchhoff.movies.databinding.FragmentPaginatedBinding
 
 abstract class PaginatedScreenFragment<Data, T : UIPaginated<Data>> : BaseFragment(R.layout.fragment_paginated) {
