+++ core/src/main/res/layout/fragment_paginated.xml
